+++ src/main/java/com/cardano/explorer/model/response/pool/projection/EpochChartProjection.java
@@ -1,7 +1,7 @@
 package com.cardano.explorer.model.response.pool.projection;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
-public interface EpochChartProjection extends BasePoolChartProjection<Integer, BigDecimal> {
+public interface EpochChartProjection extends BasePoolChartProjection<Integer, BigInteger> {
 
 }
