+++ src/main/java/com/cardano/explorer/model/response/tx/WithdrawalResponse.java
@@ -1,6 +1,6 @@
 package com.cardano.explorer.model.response.tx;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 import lombok.Builder;
 import lombok.Getter;
@@ -12,5 +12,5 @@ import lombok.Setter;
 public class WithdrawalResponse {
   private String stakeAddressFrom;
   private List<String> addressTo;
-  private BigDecimal amount;
+  private BigInteger amount;
 }
